remove ConstraintReaderMap interface
