+++ profile/src/main/java/com/scottlogic/deg/profile/guice/ProfileModule.java
@@ -20,6 +20,7 @@ import com.google.inject.AbstractModule;
 import com.scottlogic.deg.profile.reader.*;
 import com.scottlogic.deg.profile.v0_1.ProfileSchemaValidator;
 
+import java.util.Arrays;
 import java.util.stream.Stream;
 
 public class ProfileModule extends AbstractModule {
@@ -40,7 +41,7 @@ public class ProfileModule extends AbstractModule {
         bind(ProfileReader.class).to(JsonProfileReader.class);
 
         // Load built-in profile-to-constraint mappings
-        BaseConstraintReaderMap map = new BaseConstraintReaderMap(Stream.of(
+        ConstraintReaderMap map = new ConstraintReaderMap(Arrays.asList(
             new CoreAtomicTypesConstraintReaderSource(profileConfigSource.fromFilePath()),
             new FinancialTypesConstraintReaderSource(),
             new PersonalDataTypesConstraintReaderSource()
